tools/tune_viewer/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pyyaml
